web/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sse-starlette>=1.6.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
